fix(circular-deps): move GitHub field constants to clients module
Break circular dependency chain: clients → utils → config → adapters → clients
by keeping all client-related constants within the clients module itself.

Changes:
- Create src/vibe3/clients/github_field_constants.py with all GitHub field constants
- Update src/vibe3/clients/__init__.py with direct imports for proper typing
- Update src/vibe3/clients/github_issues_ops.py import source
- Batch replace GITHUB_* constant imports from vibe3.utils.constants to vibe3.clients (10 files)
- Keep non-GitHub constants (CODEAGENT_STDIN_MODE_THRESHOLD, etc.) in vibe3.utils.constants
- Remove GitHub field constants from src/vibe3/utils/constants.py

This resolves CI pytest errors:
- Circular dependencies outside L3 core found
- Module API violations (11 constants not in __all__)
