+++ src/vibe3/domain/handlers/issue_state_dispatch.py
@@ -131,7 +131,7 @@ def handle_manager_dispatch_intent(
                 state=target_state,
             )
         else:
-            from vibe3.utils.constants import GITHUB_DEFAULT_VIEW_FIELDS
+            from vibe3.clients import GITHUB_DEFAULT_VIEW_FIELDS
 
             issue_data = await loop.run_in_executor(
                 None,
