+++ src/vibe3/services/check_cleanup_service.py
@@ -370,7 +370,7 @@ class CheckCleanupService:
             from vibe3.services.label_service import LabelService
 
             gh = self._github_client or GitHubClient()
-            from vibe3.utils.constants import GITHUB_FIELDS_STATE_ONLY
+            from vibe3.clients import GITHUB_FIELDS_STATE_ONLY
 
             gh_issue = gh.view_issue(
                 issue_number, fields=list(GITHUB_FIELDS_STATE_ONLY)
